msm8960: Power HAL shouldn't control the governor
If it does, it takes away from the user's ability to control
this from user-space.

And, the Power HAL only actually controls the governor until the
user changes it.  As soon as the user changes the governor, the
sysfs for the old governor gets unloaded and the new one gets
loaded.  The only way that system_server has permission to
control the governor is because the rootdir scripts change
permissions for the governor loaded (on boot).

So, when the sysfs changes, system_server, which is running the
Power HAL, no longer has permissions to access the sysfs
attributes.

Aka: if you switch to a new governor and back, the system now
behaves completely differently because permissions have changed!

Additionally, these settings are also hurting the quality of
bluetooth audio playback for the DLX.  When the screen is off,
any increase in load on the phone causes audio artifacts because
the governor is too slow to respond to the load.

Change-Id: I92ed1896cd4db3cfae92d010e04334d2cfd0e7f2