qcom: Remove healthd extension and charger images
According to the original commit, this was needed to select
the appropriate battery capacity node on d2 and jf.

However, along the way, it seems that the custom healthd is
no longer needed, thus is no longer used on jf.

Recently, CM added its own custom healthd and charger images,
which work just fine on jf.

Hence, remove the custom healthd lib and charger images.

Change-Id: Idfa80434075818fb02eff8ba1ac33556608ed636