qcom: Remove healthd extension and charger images
According to the original commit, the healthd extension was
needed to get the correct battery capacity on d2 and jf.

However, along the way, it seems that the custom healthd is
no longer needed, is no longer used on jf.

Furthermore, CM now has its own healthd extension and charger
images, which work just fine on jf.

Hence, remove the obsolete healthd extension and charger images.

Change-Id: Idfa80434075818fb02eff8ba1ac33556608ed636