qcom: doze: Remove duplicate Ambient Display switch
Currently, we are duplicating the Ambient Display switch
in the 'Display' options.

Remove our switch for Ambient Display and use register a
ContentObserver to react to changes in Ambient Display enablement.

Also, set summaries for the gesture options whenever
Ambient Display is disabled, so that users understand why
the gesture options are disabled.

Change-Id: Iaacd27d215f18da600eb1b655df2ca06e9c46647